+++ kotlin/filament/src/jsMain/kotlin/io/github/erkko68/filament/ColorGrading.js.kt
@@ -134,6 +134,11 @@ actual class ColorGrading(internal val jsColorGrading: JSColorGrading) {
             return this
         }
 
+        actual fun fastMath(fastMath: Boolean): Builder {
+            jsBuilder.fastMath(fastMath)
+            return this
+        }
+
         actual fun build(engine: Engine): ColorGrading {
             return ColorGrading(jsBuilder.build(engine.jsEngine))
         }